requirements.txt
art
peft
pandas
pydantic
pyyaml
torch==2.0.1
transformers>=4.36.2
datasets>=2.16.0
astunparse>=1.6.3
accelerate>=0.25.0
sentencepiece>=0.1.99
deepspeed>=0.12.6
scikit-learn